refactor: share Tailcat CLI support across dial directions (#6570)
